Refactor block dialog to fix bug
