+++ ui/src/pages/namespace/Explorer/Page/poc/BlockEditor/BlockDialogProvider.tsx
@@ -1,7 +1,9 @@
 import { createContext, useContext, useState } from "react";
 
 import { AllBlocksType } from "../schema/blocks";
+import { BlockDialogContent } from "./components/BlockDialogContent";
 import { BlockPathType } from "../PageCompiler/Block";
+import { Dialog } from "~/design/Dialog";
 
 type DialogState = null | {
   action: "create" | "edit" | "delete";
@@ -25,9 +27,21 @@ export const BlockDialogProvider = ({
   children: React.ReactNode;
 }) => {
   const [dialog, setDialog] = useState<DialogState>(null);
+  /**
+   * This handler is only used for closing the dialog. For opening a dialog,
+   * we add custom onClick events to the trigger buttons.
+   */
+  const handleOnOpenChange = (open: boolean) => {
+    if (open === false) {
+      setDialog(null);
+    }
+  };
   return (
     <DialogContext.Provider value={{ dialog, setDialog }}>
-      {children}
+      <Dialog open={!!dialog} onOpenChange={handleOnOpenChange}>
+        {children}
+        <BlockDialogContent />
+      </Dialog>
     </DialogContext.Provider>
   );
 };
